pytorch 0.4.0 cleanup
