+++ attorch/train.py
@@ -1,7 +1,4 @@
 from collections import OrderedDict
-from itertools import cycle
-
-from attorch.dataset import to_variable
 
 
 def copy_state(model):
@@ -81,37 +78,3 @@ def early_stopping(model, objective, interval=5, patience=20, start=0, max_iter=
         print('Final best model! objective {:.6f}'.format(_objective(model)))
 
 
-def alternate(*args):
-    """
-    Given multiple iterators, returns a generator that alternatively visit one element from each iterator at a time.
-
-    Examples:
-        >>> list(alternate(['a', 'b', 'c'], [1, 2, 3], ['Mon', 'Tue', 'Wed']))
-        ['a', 1, 'Mon', 'b', 2, 'Tue', 'c', 3, 'Wed']
-
-    Args:
-        *args: one or more iterables (e.g. tuples, list, iterators) separated by commas
-
-    Returns:
-        A generator that alternatively visits one element at a time from the list of iterables
-    """
-    for row in zip(*args):
-        yield from row
-
-
-def cycle_datasets(trainloaders, **kwargs):
-    """
-    Cycles through datasets of train loaders.
-
-    Args:
-        trainloaders: OrderedDict with trainloaders as values
-        **kwargs: those arguments will be passed to `attorch.dataset.to_variable`
-
-    Yields:
-        readout key, input, targets
-
-    """
-    assert isinstance(trainloaders, OrderedDict), 'trainloaders must be an ordered dict'
-    for readout_key, outputs in zip(cycle(trainloaders.keys()),
-                                     to_variable(alternate(*trainloaders.values()), **kwargs)):
-        yield (readout_key,) + outputs
